pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alen-audit-si"
version = "0.2.0"
description = "Alen Audit Sistem Informasi - defensive audit & compliance reporting toolkit"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="SUPRI PES (Alen)", email="alenkesuma@gmail.com"}]
dependencies = [
  "jinja2>=3.1.2",
  "pyyaml>=6.0.1",
  "rich>=13.7.0",
  "jsonschema>=4.21.1",
]

[project.scripts]
alen-audit = "alen_audit.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
